refactored and updated skills
